+++ CHANGELOG.md
@@ -15,6 +15,7 @@ v0.36.1 (2023-09-06)
 - Restart managed components of a module loader only on if module content
   changes or the last load failed. This was specifically impacting `module.git`
   each time it pulls. (@erikbaranowski)
+- Allow overriding default `User-Agent` for `http.remote` component (@hainenber)
 
 v0.36.0 (2023-08-30)
 --------------------